fixed format string problem
git-svn-id: https://svn.code.sf.net/p/gphoto/code/trunk/libgphoto2@14620 67ed7778-7388-44ab-90cf-0a291f65f57c

+++ examples/lunkwill-canon-capture.c
@@ -28,7 +28,7 @@ static void errordumper(GPLogLevel level, const char *domain, const char *str,
   struct timeval tv;
 
   gettimeofday (&tv, NULL);
-  fprintf(stdout, "%d.%d: %s\n", tv.tv_sec, tv.tv_usec, str);
+  fprintf(stdout, "%d.%d: %s\n", (int)tv.tv_sec, (int)tv.tv_usec, str);
 }
 
 /* This seems to have no effect on where images go
